rename common to internal to prevent accident exports
