+++ packages/react/src/internal/context.ts
@@ -1,14 +1,16 @@
 import type { Client } from '@openfeature/web-sdk';
 import React from 'react';
-import type { NormalizedOptions, ReactFlagEvaluationOptions} from '../common';
-import { normalizeOptions } from '../common';
+import type { NormalizedOptions, ReactFlagEvaluationOptions } from '.';
+import { normalizeOptions } from '.';
 
 /**
  * The underlying React context.
  * DO NOT EXPORT PUBLICLY
  * @internal
  */
-export const Context = React.createContext<{ client: Client; domain?: string; options: ReactFlagEvaluationOptions } | undefined>(undefined);
+export const Context = React.createContext<
+  { client: Client; domain?: string; options: ReactFlagEvaluationOptions } | undefined
+>(undefined);
 
 /**
  * Get a normalized copy of the options used for this OpenFeatureProvider, see {@link normalizeOptions}.
