+++ packages/react/src/internal/is-equal.ts
