+++ packages/react/src/internal/options.ts
